+++ CHANGELOG.md
@@ -8,9 +8,15 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+*   `enp`:
+    *   Make `enp.testing.parametrize_xnp()` import only requested xnp modules.
 *   `ecolab`:
     *   `ecolab.inspect`: Proto are better displayed (hide attributes
         `DESCRIPTOR`, `Extensions` in sub-section)
+*   `epy`:
+    *   `epy.lazy_api_imports`: Fix infinite recursion when importing sub-module
+*   `exm`:
+    *   Add dummy implementation of the API to simplify open-sourcing.
 
 ## [1.10.0] - 2024-10-17